feat: add metadata options to es client
