+++ lib/transports/single-key-rabbit.js
@@ -4,15 +4,22 @@ const Logger = require('@pager/logger');
 
 class RabbitTransport {
 
-    constructor(publisher, logger = Logger) {
+    constructor(publisher, metdadata, logger = Logger) {
 
         this.publisher = publisher;
+        this.metdadata = metdadata;
         this.logger = logger;
     }
 
     async publish(stream, streamId, eventType, data, eventId) {
 
         this.logger.info({ msg: 'publishing event', stream, streamId, eventType, eventId });
+
+        const options = { key: `events.${stream}.created` };
+
+        if (this.metdadata) {
+            options.headers = this.metdadata;
+        }
 
         return await this.publisher.publish(
             {
@@ -21,11 +28,7 @@ class RabbitTransport {
                 streamId,
                 eventType,
                 eventId
-            },
-            {
-                key: `events.${stream}.created`
-            }
-        );
+            }, options);
     }
 }
